requirements.txt
azure-common==1.1.4
azure-nspkg==1.0.0
azure-storage==0.33.0
cffi==1.9.1
click==6.6
cryptography==1.5.3
enum34==1.1.6
Flask==0.11.1
futures==3.0.5
idna==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pkg-resources==0.0.0
pyasn1==0.1.9
pycparser==2.17
pygame==1.9.2b8
python-dateutil==2.6.0
requests==2.11.1
six==1.10.0
Werkzeug==0.11.11
